Refactor: Improve the login experience
The current module of providing one or more parameters for each login
method and overlaps in between multiple options, is too complex and
difficult to comprehend. Simplify by providing a name for each login
method and a optional dictionary of elements for the login method. This
allows an abundance of login methods without increasing the complexity
of the `__init__` method.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>